{"payload":{"allShortcutsEnabled":false,"fileTree":{"man":{"items":[{"name":"Gambler_R.Rd","path":"man/Gambler_R.Rd","contentType":"file"},{"name":"celltype_predict.Rd","path":"man/celltype_predict.Rd","contentType":"file"},{"name":"merge_perdict.Rd","path":"man/merge_perdict.Rd","contentType":"file"},{"name":"patient_group_predict.Rd","path":"man/patient_group_predict.Rd","contentType":"file"}],"totalCount":4},"":{"items":[{"name":"R","path":"R","contentType":"directory"},{"name":"data","path":"data","contentType":"directory"},{"name":"man","path":"man","contentType":"directory"},{"name":".Rbuildignore","path":".Rbuildignore","contentType":"file"},{"name":".gitignore","path":".gitignore","contentType":"file"},{"name":"DESCRIPTION","path":"DESCRIPTION","contentType":"file"},{"name":"NAMESPACE","path":"NAMESPACE","contentType":"file"},{"name":"README.md","path":"README.md","contentType":"file"},{"name":"SPECTRUM.Rproj","path":"SPECTRUM.Rproj","contentType":"file"}],"totalCount":9}},"fileTreeProcessingTime":6.478993,"foldersToFetch":[],"reducedMotionEnabled":null,"repo":{"id":743388899,"defaultBranch":"master","name":"SPECTRUM","ownerLogin":"lijinhu98","currentUserCanPush":false,"isFork":false,"isEmpty":false,"createdAt":"2024-01-15T06:02:07.000Z","ownerAvatar":"https://avatars.githubusercontent.com/u/156587536?v=4","public":true,"private":false,"isOrgOwned":false},"symbolsExpanded":false,"treeExpanded":true,"refInfo":{"name":"master","listCacheKey":"v0:1705298528.156678","canEdit":false,"refType":"branch","currentOid":"758f35fcd7c6101d15c35655805a2eec7e8b7899"},"path":"man/celltype_predict.Rd","currentUser":null,"blob":{"rawLines":["\\name{celltype_predict}","\\alias{celltype_predict}","%- Also NEED an '\\alias' for EACH other topic documented here.","\\title{","predict cell type","}","\\description{","predicted cell type","}","\\usage{","celltype_predict(matrix)","}","%- maybe also 'usage' for other objects documented here.","\\arguments{","  \\item{matrix}{","matrix Rows should be cells and columns should be genes.","}","}","","\\value{","return predicted cell type","}","","\\examples{","##---- Should be DIRECTLY executable !! ----","##-- ==>  Define data, use random,","##--\tor do  help(data=index)  for the standard data sets.","","## The function is currently defined as","function (x)","{","  }","}"],"stylingDirectives":[[{"start":1,"end":5,"cssClass":"pl-smi"},{"start":6,"end":22,"cssClass":"pl-smi"}],[{"start":1,"end":6,"cssClass":"pl-smi"},{"start":7,"end":23,"cssClass":"pl-smi"}],[{"start":1,"end":2,"cssClass":"pl-k"},{"start":3,"end":7,"cssClass":"pl-smi"},{"start":8,"end":12,"cssClass":"pl-smi"},{"start":13,"end":15,"cssClass":"pl-smi"},{"start":16,"end":24,"cssClass":"pl-s"},{"start":16,"end":17,"cssClass":"pl-pds"},{"start":17,"end":19,"cssClass":"pl-cce"},{"start":23,"end":24,"cssClass":"pl-pds"},{"start":25,"end":28,"cssClass":"pl-k"},{"start":29,"end":33,"cssClass":"pl-smi"},{"start":34,"end":39,"cssClass":"pl-smi"},{"start":40,"end":45,"cssClass":"pl-smi"},{"start":46,"end":56,"cssClass":"pl-smi"},{"start":57,"end":62,"cssClass":"pl-smi"}],[{"start":1,"end":6,"cssClass":"pl-smi"}],[{"start":0,"end":7,"cssClass":"pl-smi"},{"start":8,"end":12,"cssClass":"pl-smi"},{"start":13,"end":17,"cssClass":"pl-smi"}],[],[{"start":1,"end":12,"cssClass":"pl-smi"}],[{"start":0,"end":9,"cssClass":"pl-smi"},{"start":10,"end":14,"cssClass":"pl-smi"},{"start":15,"end":19,"cssClass":"pl-smi"}],[],[{"start":1,"end":6,"cssClass":"pl-smi"}],[{"start":17,"end":23,"cssClass":"pl-smi"}],[],[{"start":1,"end":2,"cssClass":"pl-k"},{"start":3,"end":8,"cssClass":"pl-smi"},{"start":9,"end":13,"cssClass":"pl-smi"},{"start":14,"end":21,"cssClass":"pl-s"},{"start":14,"end":15,"cssClass":"pl-pds"},{"start":20,"end":21,"cssClass":"pl-pds"},{"start":22,"end":25,"cssClass":"pl-k"},{"start":26,"end":31,"cssClass":"pl-smi"},{"start":32,"end":39,"cssClass":"pl-smi"},{"start":40,"end":50,"cssClass":"pl-smi"},{"start":51,"end":56,"cssClass":"pl-smi"}],[{"start":1,"end":10,"cssClass":"pl-smi"}],[{"start":3,"end":7,"cssClass":"pl-smi"},{"start":8,"end":14,"cssClass":"pl-smi"}],[{"start":0,"end":6,"cssClass":"pl-smi"},{"start":7,"end":11,"cssClass":"pl-smi"},{"start":12,"end":18,"cssClass":"pl-smi"},{"start":19,"end":21,"cssClass":"pl-smi"},{"start":22,"end":27,"cssClass":"pl-smi"},{"start":28,"end":31,"cssClass":"pl-smi"},{"start":32,"end":39,"cssClass":"pl-smi"},{"start":40,"end":46,"cssClass":"pl-smi"},{"start":47,"end":49,"cssClass":"pl-smi"},{"start":50,"end":56,"cssClass":"pl-smi"}],[],[],[],[{"start":1,"end":6,"cssClass":"pl-smi"}],[{"start":0,"end":6,"cssClass":"pl-k"},{"start":7,"end":16,"cssClass":"pl-smi"},{"start":17,"end":21,"cssClass":"pl-smi"},{"start":22,"end":26,"cssClass":"pl-smi"}],[],[],[{"start":1,"end":9,"cssClass":"pl-smi"}],[{"start":0,"end":44,"cssClass":"pl-c"},{"start":0,"end":1,"cssClass":"pl-c"}],[{"start":0,"end":34,"cssClass":"pl-c"},{"start":0,"end":1,"cssClass":"pl-c"}],[{"start":0,"end":57,"cssClass":"pl-c"},{"start":0,"end":1,"cssClass":"pl-c"}],[],[{"start":0,"end":39,"cssClass":"pl-c"},{"start":0,"end":1,"cssClass":"pl-c"}],[{"start":10,"end":11,"cssClass":"pl-smi"}],[],[],[],[]],"csv":null,"csvError":null,"dependabotInfo":{"showConfigurationBanner":false,"configFilePath":null,"networkDependabotPath":"/lijinhu98/SPECTRUM/network/updates","dismissConfigurationNoticePath":"/settings/dismiss-notice/dependabot_configuration_notice","configurationNoticeDismissed":null,"repoAlertsPath":"/lijinhu98/SPECTRUM/security/dependabot","repoSecurityAndAnalysisPath":"/lijinhu98/SPECTRUM/settings/security_analysis","repoOwnerIsOrg":false,"currentUserCanAdminRepo":false},"displayName":"celltype_predict.Rd","displayUrl":"https://github.com/lijinhu98/SPECTRUM/blob/master/man/celltype_predict.Rd?raw=true","headerInfo":{"blobSize":"609 Bytes","deleteInfo":{"deleteTooltip":"You must be signed in to make or propose changes"},"editInfo":{"editTooltip":"You must be signed in to make or propose changes"},"ghDesktopPath":"https://desktop.github.com","gitLfsPath":null,"onBranch":true,"shortPath":"d433c6e","siteNavLoginPath":"/login?return_to=https%3A%2F%2Fgithub.com%2Flijinhu98%2FSPECTRUM%2Fblob%2Fmaster%2Fman%2Fcelltype_predict.Rd","isCSV":false,"isRichtext":false,"toc":null,"lineInfo":{"truncatedLoc":"34","truncatedSloc":"30"},"mode":"file"},"image":false,"isCodeownersFile":null,"isPlain":false,"isValidLegacyIssueTemplate":false,"issueTemplateHelpUrl":"https://docs.github.com/articles/about-issue-and-pull-request-templates","issueTemplate":null,"discussionTemplate":null,"language":"R","languageID":307,"large":false,"loggedIn":false,"newDiscussionPath":"/lijinhu98/SPECTRUM/discussions/new","newIssuePath":"/lijinhu98/SPECTRUM/issues/new","planSupportInfo":{"repoIsFork":null,"repoOwnedByCurrentUser":null,"requestFullPath":"/lijinhu98/SPECTRUM/blob/master/man/celltype_predict.Rd","showFreeOrgGatedFeatureMessage":null,"showPlanSupportBanner":null,"upgradeDataAttributes":null,"upgradePath":null},"publishBannersInfo":{"dismissActionNoticePath":"/settings/dismiss-notice/publish_action_from_dockerfile","dismissStackNoticePath":"/settings/dismiss-notice/publish_stack_from_file","releasePath":"/lijinhu98/SPECTRUM/releases/new?marketplace=true","showPublishActionBanner":false,"showPublishStackBanner":false},"rawBlobUrl":"https://github.com/lijinhu98/SPECTRUM/raw/master/man/celltype_predict.Rd","renderImageOrRaw":false,"richText":null,"renderedFileInfo":null,"shortPath":null,"tabSize":8,"topBannersInfo":{"overridingGlobalFundingFile":false,"globalPreferredFundingPath":null,"repoOwner":"lijinhu98","repoName":"SPECTRUM","showInvalidCitationWarning":false,"citationHelpUrl":"https://docs.github.com/en/github/creating-cloning-and-archiving-repositories/creating-a-repository-on-github/about-citation-files","showDependabotConfigurationBanner":false,"actionsOnboardingTip":null},"truncated":false,"viewable":true,"workflowRedirectUrl":null,"symbols":{"timed_out":false,"not_analyzed":true,"symbols":[]}},"copilotInfo":null,"copilotAccessAllowed":false,"csrf_tokens":{"/lijinhu98/SPECTRUM/branches":{"post":"DWoEqciuUcdQ1vbqHDHl9BzWWrXuO1oAmXvsnYX8UiQyCtL-RqRS0PfSAEiMkPriUNCxLDOxXkx_cncwy-gSzA"},"/repos/preferences":{"post":"4KVRY-JQyS10DEUIv_CLiwwp8ORej_q4ViBb9nEr_gzNCYDSaM-fxPvZn4cpKvywxl6XQa9q7P-jjT0lZFOTQw"}}},"title":"SPECTRUM/man/celltype_predict.Rd at master · lijinhu98/SPECTRUM"}