{"payload":{"allShortcutsEnabled":false,"fileTree":{"man":{"items":[{"name":"Gambler_R.Rd","path":"man/Gambler_R.Rd","contentType":"file"},{"name":"celltype_predict.Rd","path":"man/celltype_predict.Rd","contentType":"file"},{"name":"merge_perdict.Rd","path":"man/merge_perdict.Rd","contentType":"file"},{"name":"patient_group_predict.Rd","path":"man/patient_group_predict.Rd","contentType":"file"}],"totalCount":4},"":{"items":[{"name":"R","path":"R","contentType":"directory"},{"name":"data","path":"data","contentType":"directory"},{"name":"man","path":"man","contentType":"directory"},{"name":".Rbuildignore","path":".Rbuildignore","contentType":"file"},{"name":".gitignore","path":".gitignore","contentType":"file"},{"name":"DESCRIPTION","path":"DESCRIPTION","contentType":"file"},{"name":"NAMESPACE","path":"NAMESPACE","contentType":"file"},{"name":"README.md","path":"README.md","contentType":"file"},{"name":"SPECTRUM.Rproj","path":"SPECTRUM.Rproj","contentType":"file"}],"totalCount":9}},"fileTreeProcessingTime":6.406359,"foldersToFetch":[],"reducedMotionEnabled":null,"repo":{"id":743388899,"defaultBranch":"master","name":"SPECTRUM","ownerLogin":"lijinhu98","currentUserCanPush":false,"isFork":false,"isEmpty":false,"createdAt":"2024-01-15T06:02:07.000Z","ownerAvatar":"https://avatars.githubusercontent.com/u/156587536?v=4","public":true,"private":false,"isOrgOwned":false},"symbolsExpanded":false,"treeExpanded":true,"refInfo":{"name":"master","listCacheKey":"v0:1705298528.156678","canEdit":false,"refType":"branch","currentOid":"758f35fcd7c6101d15c35655805a2eec7e8b7899"},"path":"man/Gambler_R.Rd","currentUser":null,"blob":{"rawLines":["\\name{Gambler_R}","\\alias{Gambler_R}","%- Also NEED an '\\alias' for EACH other topic documented here.","\\title{","predict","}","\\description{","predict","}","\\usage{","Gambler_R(test_r, prob_r,gene_list = NULL)","}","%- maybe also 'usage' for other objects documented here.","\\arguments{","  \\item{test_r}{","  Rows should be cells and columns should be genes.","}","\\item{prob_r}{","  trained scibet model  Rows should be genes and columns should be cell types.The genes must be matched with test_r","}","\\item{gene_list}{","  a list of genes","}","}","","\\value{","return the predict","}","","\\examples{","##---- Should be DIRECTLY executable !! ----","##-- ==>  Define data, use random,","##--\tor do  help(data=index)  for the standard data sets.","","## The function is currently defined as","function (x)","{","  }","}"],"stylingDirectives":[[{"start":1,"end":5,"cssClass":"pl-smi"},{"start":6,"end":15,"cssClass":"pl-smi"}],[{"start":1,"end":6,"cssClass":"pl-smi"},{"start":7,"end":16,"cssClass":"pl-smi"}],[{"start":1,"end":2,"cssClass":"pl-k"},{"start":3,"end":7,"cssClass":"pl-smi"},{"start":8,"end":12,"cssClass":"pl-smi"},{"start":13,"end":15,"cssClass":"pl-smi"},{"start":16,"end":24,"cssClass":"pl-s"},{"start":16,"end":17,"cssClass":"pl-pds"},{"start":17,"end":19,"cssClass":"pl-cce"},{"start":23,"end":24,"cssClass":"pl-pds"},{"start":25,"end":28,"cssClass":"pl-k"},{"start":29,"end":33,"cssClass":"pl-smi"},{"start":34,"end":39,"cssClass":"pl-smi"},{"start":40,"end":45,"cssClass":"pl-smi"},{"start":46,"end":56,"cssClass":"pl-smi"},{"start":57,"end":62,"cssClass":"pl-smi"}],[{"start":1,"end":6,"cssClass":"pl-smi"}],[{"start":0,"end":7,"cssClass":"pl-smi"}],[],[{"start":1,"end":12,"cssClass":"pl-smi"}],[{"start":0,"end":7,"cssClass":"pl-smi"}],[],[{"start":1,"end":6,"cssClass":"pl-smi"}],[{"start":10,"end":16,"cssClass":"pl-smi"},{"start":18,"end":24,"cssClass":"pl-smi"},{"start":25,"end":34,"cssClass":"pl-v"},{"start":35,"end":36,"cssClass":"pl-k"},{"start":37,"end":41,"cssClass":"pl-c1"}],[],[{"start":1,"end":2,"cssClass":"pl-k"},{"start":3,"end":8,"cssClass":"pl-smi"},{"start":9,"end":13,"cssClass":"pl-smi"},{"start":14,"end":21,"cssClass":"pl-s"},{"start":14,"end":15,"cssClass":"pl-pds"},{"start":20,"end":21,"cssClass":"pl-pds"},{"start":22,"end":25,"cssClass":"pl-k"},{"start":26,"end":31,"cssClass":"pl-smi"},{"start":32,"end":39,"cssClass":"pl-smi"},{"start":40,"end":50,"cssClass":"pl-smi"},{"start":51,"end":56,"cssClass":"pl-smi"}],[{"start":1,"end":10,"cssClass":"pl-smi"}],[{"start":3,"end":7,"cssClass":"pl-smi"},{"start":8,"end":14,"cssClass":"pl-smi"}],[{"start":2,"end":6,"cssClass":"pl-smi"},{"start":7,"end":13,"cssClass":"pl-smi"},{"start":14,"end":16,"cssClass":"pl-smi"},{"start":17,"end":22,"cssClass":"pl-smi"},{"start":23,"end":26,"cssClass":"pl-smi"},{"start":27,"end":34,"cssClass":"pl-smi"},{"start":35,"end":41,"cssClass":"pl-smi"},{"start":42,"end":44,"cssClass":"pl-smi"},{"start":45,"end":51,"cssClass":"pl-smi"}],[],[{"start":1,"end":5,"cssClass":"pl-smi"},{"start":6,"end":12,"cssClass":"pl-smi"}],[{"start":2,"end":9,"cssClass":"pl-smi"},{"start":10,"end":16,"cssClass":"pl-smi"},{"start":17,"end":22,"cssClass":"pl-smi"},{"start":24,"end":28,"cssClass":"pl-smi"},{"start":29,"end":35,"cssClass":"pl-smi"},{"start":36,"end":38,"cssClass":"pl-smi"},{"start":39,"end":44,"cssClass":"pl-smi"},{"start":45,"end":48,"cssClass":"pl-smi"},{"start":49,"end":56,"cssClass":"pl-smi"},{"start":57,"end":63,"cssClass":"pl-smi"},{"start":64,"end":66,"cssClass":"pl-smi"},{"start":67,"end":71,"cssClass":"pl-smi"},{"start":72,"end":81,"cssClass":"pl-smi"},{"start":82,"end":87,"cssClass":"pl-smi"},{"start":88,"end":92,"cssClass":"pl-smi"},{"start":93,"end":95,"cssClass":"pl-smi"},{"start":96,"end":103,"cssClass":"pl-smi"},{"start":104,"end":108,"cssClass":"pl-smi"},{"start":109,"end":115,"cssClass":"pl-smi"}],[],[{"start":1,"end":5,"cssClass":"pl-smi"},{"start":6,"end":15,"cssClass":"pl-smi"}],[{"start":2,"end":3,"cssClass":"pl-smi"},{"start":4,"end":8,"cssClass":"pl-smi"},{"start":9,"end":11,"cssClass":"pl-smi"},{"start":12,"end":17,"cssClass":"pl-smi"}],[],[],[],[{"start":1,"end":6,"cssClass":"pl-smi"}],[{"start":0,"end":6,"cssClass":"pl-k"},{"start":7,"end":10,"cssClass":"pl-smi"},{"start":11,"end":18,"cssClass":"pl-smi"}],[],[],[{"start":1,"end":9,"cssClass":"pl-smi"}],[{"start":0,"end":44,"cssClass":"pl-c"},{"start":0,"end":1,"cssClass":"pl-c"}],[{"start":0,"end":34,"cssClass":"pl-c"},{"start":0,"end":1,"cssClass":"pl-c"}],[{"start":0,"end":57,"cssClass":"pl-c"},{"start":0,"end":1,"cssClass":"pl-c"}],[],[{"start":0,"end":39,"cssClass":"pl-c"},{"start":0,"end":1,"cssClass":"pl-c"}],[{"start":10,"end":11,"cssClass":"pl-smi"}],[],[],[],[]],"csv":null,"csvError":null,"dependabotInfo":{"showConfigurationBanner":false,"configFilePath":null,"networkDependabotPath":"/lijinhu98/SPECTRUM/network/updates","dismissConfigurationNoticePath":"/settings/dismiss-notice/dependabot_configuration_notice","configurationNoticeDismissed":null,"repoAlertsPath":"/lijinhu98/SPECTRUM/security/dependabot","repoSecurityAndAnalysisPath":"/lijinhu98/SPECTRUM/settings/security_analysis","repoOwnerIsOrg":false,"currentUserCanAdminRepo":false},"displayName":"Gambler_R.Rd","displayUrl":"https://github.com/lijinhu98/SPECTRUM/blob/master/man/Gambler_R.Rd?raw=true","headerInfo":{"blobSize":"749 Bytes","deleteInfo":{"deleteTooltip":"You must be signed in to make or propose changes"},"editInfo":{"editTooltip":"You must be signed in to make or propose changes"},"ghDesktopPath":"https://desktop.github.com","gitLfsPath":null,"onBranch":true,"shortPath":"21aefa0","siteNavLoginPath":"/login?return_to=https%3A%2F%2Fgithub.com%2Flijinhu98%2FSPECTRUM%2Fblob%2Fmaster%2Fman%2FGambler_R.Rd","isCSV":false,"isRichtext":false,"toc":null,"lineInfo":{"truncatedLoc":"40","truncatedSloc":"36"},"mode":"file"},"image":false,"isCodeownersFile":null,"isPlain":false,"isValidLegacyIssueTemplate":false,"issueTemplateHelpUrl":"https://docs.github.com/articles/about-issue-and-pull-request-templates","issueTemplate":null,"discussionTemplate":null,"language":"R","languageID":307,"large":false,"loggedIn":false,"newDiscussionPath":"/lijinhu98/SPECTRUM/discussions/new","newIssuePath":"/lijinhu98/SPECTRUM/issues/new","planSupportInfo":{"repoIsFork":null,"repoOwnedByCurrentUser":null,"requestFullPath":"/lijinhu98/SPECTRUM/blob/master/man/Gambler_R.Rd","showFreeOrgGatedFeatureMessage":null,"showPlanSupportBanner":null,"upgradeDataAttributes":null,"upgradePath":null},"publishBannersInfo":{"dismissActionNoticePath":"/settings/dismiss-notice/publish_action_from_dockerfile","dismissStackNoticePath":"/settings/dismiss-notice/publish_stack_from_file","releasePath":"/lijinhu98/SPECTRUM/releases/new?marketplace=true","showPublishActionBanner":false,"showPublishStackBanner":false},"rawBlobUrl":"https://github.com/lijinhu98/SPECTRUM/raw/master/man/Gambler_R.Rd","renderImageOrRaw":false,"richText":null,"renderedFileInfo":null,"shortPath":null,"tabSize":8,"topBannersInfo":{"overridingGlobalFundingFile":false,"globalPreferredFundingPath":null,"repoOwner":"lijinhu98","repoName":"SPECTRUM","showInvalidCitationWarning":false,"citationHelpUrl":"https://docs.github.com/en/github/creating-cloning-and-archiving-repositories/creating-a-repository-on-github/about-citation-files","showDependabotConfigurationBanner":false,"actionsOnboardingTip":null},"truncated":false,"viewable":true,"workflowRedirectUrl":null,"symbols":{"timed_out":false,"not_analyzed":true,"symbols":[]}},"copilotInfo":null,"copilotAccessAllowed":false,"csrf_tokens":{"/lijinhu98/SPECTRUM/branches":{"post":"RFaf4q9iR4A9fzxLvau5Db2mW7cx_6DUYGnWnszCna3yY-VB4a5ZNF9NZ_Iu0zUW1yoqJx3mAJ_pN9OMDH_XSA"},"/repos/preferences":{"post":"6KIyQP4GAAu3CZEp6yWn6O_46H9CGHs91BJ-_0Aw1HvOvsGjmra543s07QBhhWItsnuF4peEHzUoJVtbQGKJRg"}}},"title":"SPECTRUM/man/Gambler_R.Rd at master · lijinhu98/SPECTRUM"}